+++ src/main/java/level/up/kaabar/Drinks/MyDrinkSevrvice.java
@@ -1,13 +1,13 @@
 package level.up.kaabar.Drinks;
 
-import level.up.kaabar.config.model.Drink;
+import level.up.kaabar.model.Drink;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 
 @Component
 public interface MyDrinkSevrvice {
-    public Drink newDrinkAdd (String drinkName, String brand, double price, int quantity, String typ);
+    public Drink newDrinkAdd (String drinkName, String brand, int price, int quantity, String typ);
 
     public Drink findByID (int id);
